backend/requirements.txt
flask
flask-cors
pymupdf
opencv-python-headless
pytesseract
pillow
lxml
numpy
python-dotenv
gunicorn